+++ constants/role.constant.js
@@ -38,6 +38,10 @@ const hackerRole = {
         Constants.Routes.hackerRoutes.patchSelfConfirmationById,
         Constants.Routes.hackerRoutes.getSelf,
 
+        Constants.Routes.travelRoutes.getSelf,
+        Constants.Routes.travelRoutes.getSelfById,
+        Constants.Routes.travelRoutes.getSelfByEmail,
+
         Constants.Routes.teamRoutes.join,
         Constants.Routes.teamRoutes.patchSelfById,
         Constants.Routes.teamRoutes.post,