+++ constants/success.constant.js
@@ -27,6 +27,10 @@ const HACKER_SENT_DAY_OF = "Hacker day-of email sent.";
 const RESUME_UPLOAD = "Resume upload successful.";
 const RESUME_DOWNLOAD = "Resume download successful.";
 
+const TRAVEL_READ = "Travel retrieval successful.";
+const TRAVEL_CREATE = "Travel creation successful.";
+const TRAVEL_UPDATE = "Travel update successful.";
+
 const ROLE_CREATE = "Role creation successful.";
 
 const SEARCH_QUERY = "Query search successful. Returning results.";
@@ -77,6 +81,10 @@ module.exports = {
 
     RESUME_UPLOAD: RESUME_UPLOAD,
     RESUME_DOWNLOAD: RESUME_DOWNLOAD,
+
+    TRAVEL_READ: TRAVEL_READ,
+    TRAVEL_CREATE: TRAVEL_CREATE,
+    TRAVE_UPDATE: TRAVEL_UPDATE,
 
     ROLE_CREATE: ROLE_CREATE,