+++ middlewares/hacker.middleware.js
@@ -7,6 +7,7 @@ const Services = {
     Storage: require("../services/storage.service"),
     Email: require("../services/email.service"),
     Account: require("../services/account.service"),
+    Travel: require("../services/travel.service"),
     Env: require("../services/env.service")
 };
 const Middleware = {
@@ -195,7 +196,7 @@ async function validateConfirmedStatusFromArrayofHackerIds(req, res, next) {
         return next({
             status: 404,
             message: Constants.Error.HACKER_404_MESSAGE
-        })
+        });
     }
     const promise = new Promise((resolve, reject) => {
         req.body.ids.forEach(async (id, index) => {
@@ -203,7 +204,7 @@ async function validateConfirmedStatusFromArrayofHackerIds(req, res, next) {
             if (hacker == null) {
                 req.body.errors.push([id, "Hacker Not Found"]);
                 req.body.ids.splice(index, 1);
-                if (index == req.body.ids.length-1) resolve();
+                if (index == req.body.ids.length - 1) resolve();
                 return;
             }
             const account = await Services.Account.findById(hacker.accountId);
@@ -213,7 +214,7 @@ async function validateConfirmedStatusFromArrayofHackerIds(req, res, next) {
                 console.log(req.body.errors);
                 req.body.ids.splice(index, 1);
             }
-            if (index == req.body.ids.length-1) resolve();
+            if (index == req.body.ids.length - 1) resolve();
         });
     });
 
@@ -395,19 +396,27 @@ async function sendStatusUpdateEmailForMultipleIds(req, res, next) {
         return next();
     } else {
         // send it to the hacker that is being updated.
-        const promise = new Promise ((resolve, reject) => {
+        const promise = new Promise((resolve, reject) => {
             req.body.ids.forEach(async (id, index) => {
                 const hacker = await Services.Hacker.findById(id);
-                const account = await Services.Account.findById(hacker.accountId);
+                const account = await Services.Account.findById(
+                    hacker.accountId
+                );
                 if (!hacker) {
-                    req.body.errors.push(id, Constants.Error.HACKER_404_MESSAGE);
+                    req.body.errors.push(
+                        id,
+                        Constants.Error.HACKER_404_MESSAGE
+                    );
                     req.body.ids.splice(index, 1);
-                    if (index == req.body.ids.length-1) resolve();
+                    if (index == req.body.ids.length - 1) resolve();
                     return;
                 } else if (!account) {
-                    req.body.errors.push(id, Constants.Error.GENERIC_500_MESSAGE);
+                    req.body.errors.push(
+                        id,
+                        Constants.Error.GENERIC_500_MESSAGE
+                    );
                     req.body.ids.splice(index, 1);
-                    if (index == req.body.ids.length-1) resolve();
+                    if (index == req.body.ids.length - 1) resolve();
                     return;
                 }
 
@@ -417,13 +426,46 @@ async function sendStatusUpdateEmailForMultipleIds(req, res, next) {
                     req.body.status,
                     next
                 );
-                if (index == req.body.ids.length-1) resolve();
-            })
-        })
+                if (index == req.body.ids.length - 1) resolve();
+            });
+        });
 
         promise.then(() => {
             return next();
-        })
+        });
+    }
+}
+/**
+ * Sends a preset email to a user if a status change occured with email params.
+ * @param {{body: {status?: string}, params: {email: string}}} req
+ * @param {*} res
+ * @param {(err?:*)=>void} next
+ */
+async function completeStatusUpdateEmail(req, res, next) {
+    //skip if the status doesn't exist
+    if (!req.body.hacker.status) {
+        return next();
+    } else {
+        // send it to the hacker that is being updated.
+        const hacker = await Services.Hacker.findById(req.body.hacker._id);
+        const account = await Services.Account.findById(hacker.accountId);
+        if (!hacker) {
+            return next({
+                status: 404,
+                message: Constants.Error.HACKER_404_MESSAGE
+            });
+        } else if (!account) {
+            return next({
+                status: 500,
+                message: Constants.Error.GENERIC_500_MESSAGE
+            });
+        }
+        Services.Email.sendStatusUpdate(
+            account.firstName,
+            account.email,
+            req.body.hacker.status,
+            next
+        );
     }
 }
 
@@ -610,6 +652,48 @@ async function updateHacker(req, res, next) {
             });
         }
         req.email = acct.email;
+
+        // If this hacker has a travel account associated with it, then update request to reflect amount wanted for travel
+        const travel = await Services.Travel.findByHackerId(hacker.id);
+        if (travel) {
+            await Services.Travel.updateOne(travel.id, {
+                request: hacker.application.accommodation.travel
+            });
+        }
+        return next();
+    } else {
+        return next({
+            status: 404,
+            message: Constants.Error.HACKER_404_MESSAGE,
+            data: {
+                id: req.params.id
+            }
+        });
+    }
+}
+
+/**
+ * Updates a hacker that is specified by req.body.hacker._id, and then sets req.email
+ * to the email of the hacker, found in Account.
+ * @param {{params:{_id: string}, body: *}} req
+ * @param {*} res
+ * @param {*} next
+ */
+async function obtainEmailByHackerId(req, res, next) {
+    const hacker = await Services.Hacker.findById(req.body.hacker._id);
+    if (hacker) {
+        const acct = await Services.Account.findById(hacker.accountId);
+        if (!acct) {
+            return next({
+                status: 500,
+                message: Constants.Error.HACKER_UPDATE_500_MESSAGE,
+                data: {
+                    hackerId: hacker.id,
+                    accountId: hacker.accountId
+                }
+            });
+        }
+        req.email = acct.email;
         return next();
     } else {
         return next({
@@ -632,24 +716,27 @@ async function updateHacker(req, res, next) {
 async function updateBatchHacker(req, res, next) {
     const promise = new Promise((resolve, reject) => {
         req.body.ids.forEach(async (id, index) => {
-        const hacker = await Services.Hacker.updateOne(id, req.body);
-        if (hacker) {
-            const acct = await Services.Account.findById(hacker.accountId);
-            if (!acct) {
-                req.body.errors.push([id, Constants.Error.HACKER_UPDATE_500_MESSAGE]);
+            const hacker = await Services.Hacker.updateOne(id, req.body);
+            if (hacker) {
+                const acct = await Services.Account.findById(hacker.accountId);
+                if (!acct) {
+                    req.body.errors.push([
+                        id,
+                        Constants.Error.HACKER_UPDATE_500_MESSAGE
+                    ]);
+                    req.body.ids.splice(index, 1);
+                    if (index == req.body.ids.length - 1) resolve();
+                    return;
+                }
+                req.email = acct.email;
+                if (index == req.body.ids.length - 1) resolve();
+            } else {
+                req.body.errors([id, Constants.Error.HACKER_404_MESSAGE]);
                 req.body.ids.splice(index, 1);
-                if (index == req.body.ids.length-1) resolve();
-                return;
+                if (index == req.body.ids.length - 1) resolve();
             }
-            req.email = acct.email;
-            if (index == req.body.ids.length-1) resolve();
-        } else {
-            req.body.errors([id, Constants.Error.HACKER_404_MESSAGE]);
-            req.body.ids.splice(index, 1);
-            if (index == req.body.ids.length-1) resolve();
-        }
         });
-    })
+    });
     promise.then(() => {
         return next();
     });
@@ -663,6 +750,19 @@ async function updateBatchHacker(req, res, next) {
  */
 function parseAccept(req, res, next) {
     req.body.status = Constants.General.HACKER_STATUS_ACCEPTED;
+    req.hackerId = req.params.id;
+    next();
+}
+
+/**
+ * Sets req.body.hacker.status to Accepted for next middleware.
+ * @param {{params:{email: string}, body: *}} req
+ * @param {*} res
+ * @param {*} next
+ */
+function parseAcceptEmail(req, res, next) {
+    req.body.hacker.status = Constants.General.HACKER_STATUS_ACCEPTED;
+    req.hackerId = req.body.hacker._id;
     next();
 }
 
@@ -678,7 +778,7 @@ function parseBatchAccept(req, res, next) {
 }
 
 /**
- * @function createhacker
+ * @function createHacker
  * @param {{body: {hackerDetails: object}}} req
  * @param {*} res
  * @param {(err?)=>void} next
@@ -799,6 +899,7 @@ module.exports = {
     updateHacker: Middleware.Util.asyncMiddleware(updateHacker),
     updateBatchHacker: Middleware.Util.asyncMiddleware(updateBatchHacker),
     parseAccept: parseAccept,
+    parseAcceptEmail: parseAcceptEmail,
     parseBatch: parseBatchAccept,
     validateConfirmedStatusFromAccountId: Middleware.Util.asyncMiddleware(
         validateConfirmedStatusFromAccountId
@@ -811,6 +912,9 @@ module.exports = {
     ),
     validateConfirmedStatusFromObject: Middleware.Util.asyncMiddleware(
         validateConfirmedStatusFromObject
+    ),
+    completeStatusUpdateEmail: Middleware.Util.asyncMiddleware(
+        completeStatusUpdateEmail
     ),
     checkDuplicateAccountLinks: Middleware.Util.asyncMiddleware(
         checkDuplicateAccountLinks