+++ .github/CHANGELOG.md
@@ -5,6 +5,49 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0](https://github.com/hackmcgill/hackerapi/tree/2.5.0) - 2020-04-17
+
+### Added
+
+-   Add batch accept script
+
+## [2.4.0](https://github.com/hackmcgill/hackerapi/tree/2.4.0) - 2020-01-31
+
+### Added
+
+-   Added the live site link to the check in email
+
+## [2.3.1](https://github.com/hackmcgill/hackerapi/tree/2.3.1) - 2020-01-15
+
+### Removed
+
+-   Removed travel routes from hacker
+
+## [2.3.0](https://github.com/hackmcgill/hackerapi/tree/2.3.0) - 2020-01-14
+
+### Added
+
+-   Added travel model
+-   Added route to create travel
+-   Added routes to look up travel by id, email or self
+-   Added routes to status or offer of an existing travel
+
+### Changed
+
+-   Update hacker status emails: accepted, checked-in, confirmed, declined, waitlisted, withdrawn
+
+## [2.2.0](https://github.com/hackmcgill/hackerapi/tree/2.2.0) - 2020-01-12
+
+### Added
+
+-   Add route to accept hacker by email
+
+## [2.1.3](https://github.com/hackmcgill/hackerapi/tree/2.1.3) - 2020-01-11
+
+### Changed
+
+-   Comment out application close validation
+
 ## [2.1.2](https://github.com/hackmcgill/hackerapi/tree/2.1.2) - 2020-01-05
 
 ### Changed